// api/generated/category-api/api.ts
export interface CategoryResponse {
    /**
     * 
     * @type {string}
     * @memberof CategoryResponse
     */
    'id': string;
    /**
     * 
     * @type {number}
     * @memberof CategoryResponse
     */
    'version': number;
    /**
     * 
     * @type {string}
     * @memberof CategoryResponse
     */
    'name': string;
    /**
     * 
     * @type {boolean}
     * @memberof CategoryResponse
     */
    'enabled': boolean;
    /**
     * 
     * @type {string}
     * @memberof CategoryResponse
     */
    'createdAt': string;
    /**
     * 
     * @type {string}
     * @memberof CategoryResponse
     */
    'modifiedAt': string;
}
/**
 * 
 * @export
 * @interface CreateCategoryRequest
 */
export interface CreateCategoryRequest {
    /**
     * 
     * @type {string}
     * @memberof CreateCategoryRequest
     */
    'name': string;
    /**
     * 
     * @type {boolean}
     * @memberof CreateCategoryRequest
     */
    'enabled': boolean;
}
// ...
export interface UpdateCategoryRequest {
    /**
     * 
     * @type {string}
     * @memberof UpdateCategoryRequest
     */
    'id': string;
    /**
     * 
     * @type {string}
     * @memberof UpdateCategoryRequest
     */
    'name': string;
    /**
     * 
     * @type {boolean}
     * @memberof UpdateCategoryRequest
     */
    'enabled': boolean;
    /**
     * 
     * @type {number}
     * @memberof UpdateCategoryRequest
     */
    'version': number;
}
// ...
export class DefaultApi extends BaseAPI {
    /**
     * 
     * @summary Create a new category
     * @param {CreateCategoryRequest} createCategoryRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DefaultApi
     */
    public createCategory(createCategoryRequest: CreateCategoryRequest, options?: RawAxiosRequestConfig) {
        return DefaultApiFp(this.configuration).createCategory(createCategoryRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Get a list of all categories
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DefaultApi
     */
    public getAll(options?: RawAxiosRequestConfig) {
        return DefaultApiFp(this.configuration).getAll(options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Get a category by ID
     * @param {string} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DefaultApi
     */
    public getCategoryById(id: string, options?: RawAxiosRequestConfig) {
        return DefaultApiFp(this.configuration).getCategoryById(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Update an existing category
     * @param {UpdateCategoryRequest} updateCategoryRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DefaultApi
     */
    public updateCategory(updateCategoryRequest: UpdateCategoryRequest, options?: RawAxiosRequestConfig) {
        return DefaultApiFp(this.configuration).updateCategory(updateCategoryRequest, options).then((request) => request(this.axios, this.basePath));
    }
}